+++ lib/aliyun_oss/client.ex
@@ -18,6 +18,9 @@ defmodule Aliyun.Oss.Client do
       {:ok, %Req.Response{status: status, body: body}} ->
         {:error, Error.parse(%Error{body: body, status_code: status})}
 
+      {:error, %Jason.DecodeError{data: data}} ->
+        {:error, %Error{body: data, status_code: 900, parsed_details: %{"message" => "JSON decode error"}}}
+
       {:error, %{reason: reason}} ->
         {:error, reason}
     end