+++ lib/aliyun_oss/object.ex
@@ -12,8 +12,7 @@ defmodule Aliyun.Oss.Object do
 
   """
 
-  alias Aliyun.Oss.Config
-  alias Aliyun.Oss.Service
+  alias Aliyun.Oss.{Config, Service, Sign}
   alias Aliyun.Oss.Client.{Request, Response, Error}
 
   @type error() ::
@@ -23,6 +22,10 @@ defmodule Aliyun.Oss.Object do
   HeadObject - gets the metadata of an object.
 
   The content of the object is not returned.
+
+  ## Options
+
+  - `:headers` - Defaults to `%{}`
 
   ## Examples
 
@@ -53,10 +56,10 @@ defmodule Aliyun.Oss.Object do
       {:error, %Aliyun.Oss.Client.Error{status_code: 404, body: "", parsed_details: nil}}
 
   """
-  @spec head_object(Config.t(), String.t(), String.t(), map()) ::
+  @spec head_object(Config.t(), String.t(), String.t(), keyword()) ::
           {:error, error()} | {:ok, Response.t()}
-  def head_object(%Config{} = config, bucket, object, headers \\ %{}) do
-    Service.head(config, bucket, object, headers: headers)
+  def head_object(%Config{} = config, bucket, object, options \\ []) do
+    Service.head(config, bucket, object, options)
   end
 
   @doc """
@@ -94,6 +97,11 @@ defmodule Aliyun.Oss.Object do
   @doc """
   GetObject - gets an object.
 
+  ## Options
+
+  - `:query_params` - Defaults to `%{}`
+  - `:headers` - Defaults to `%{}`
+
   ## Examples
 
       iex> Aliyun.Oss.Object.get_object(config, "some-bucket", "some-object")
@@ -119,15 +127,37 @@ defmodule Aliyun.Oss.Object do
           }
         }
       }
-      iex> Aliyun.Oss.Object.get_object(config, "some-bucket", "some-object", %{}, %{"response-cache-control" => "no-cache"})
+      iex> Aliyun.Oss.Object.get_object(config, "some-bucket", "some-object", headers: %{}, query_params: %{"response-cache-control" => "no-cache"})
       {:ok, %Aliyun.Oss.Client.Response{
           data: <<208, 207, 17, 224, 161, 177, 26, 225, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
             0, 0, 0, 62, 0, 3, 0, 254, 255, 9, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18,
             0, 0, 0, ...>>,
           headers: %{
             "accept-ranges" => ["bytes"],
-            "connection" => ["keep-alive"],
             "cache-control" => ["no-cache],
+            "connection" => ["keep-alive"],
+            "content-md5" => ["wpqajJtzJSpf8lOY/W4Hqg=="],
+            "content-type" => ["text/plain"],
+            "date" => ["Fri, 04 Jul 2025 05:45:44 GMT"],
+            "etag" => ["\"C29000000000000000000000000000AA\""],
+            "last-modified" => ["Fri, 15 Jan 2021 09:16:13 GMT"],
+            "server" => ["AliyunOSS"],
+            "x-oss-hash-crc64ecma" => ["587015626014620604"],
+            "x-oss-object-type" => ["Normal"],
+            "x-oss-request-id" => ["6860000000000000000000A3"],
+            "x-oss-server-time" => ["41"],
+            "x-oss-storage-class" => ["Standard"],
+            "x-oss-version-id" => ["null"]
+          }
+        }
+      }
+      iex> Aliyun.Oss.Object.get_object(config, "some-bucket", "some-object", headers: %{"range" => "bytes=0-2"}, query_params: %{"response-cache-control" => "no-cache"})
+      {:ok, %Aliyun.Oss.Client.Response{
+          data: "abc",
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "cache-control" => ["no-cache],
+            "connection" => ["keep-alive"],
             "content-md5" => ["wpqajJtzJSpf8lOY/W4Hqg=="],
             "content-type" => ["text/plain"],
             "date" => ["Fri, 04 Jul 2025 05:45:44 GMT"],
@@ -145,10 +175,10 @@ defmodule Aliyun.Oss.Object do
       }
 
   """
-  @spec get_object(Config.t(), String.t(), String.t(), map(), map()) ::
+  @spec get_object(Config.t(), String.t(), String.t(), keyword()) ::
           {:error, error()} | {:ok, Response.t()}
-  def get_object(%Config{} = config, bucket, object, headers \\ %{}, query_params \\ %{}) do
-    Service.get(config, bucket, object, headers: headers, query_params: query_params)
+  def get_object(%Config{} = config, bucket, object, options \\ []) do
+    Service.get(config, bucket, object, options)
   end
 
   @doc """
@@ -164,25 +194,24 @@ defmodule Aliyun.Oss.Object do
 
       iex> select_request = %{
         "SelectRequest" => %{
-          "Expression" => "c2VsZWN0ICogZnJvbSBvc3NvYmplY3QuY29udGFjdHNbKl0gcyB3aGVyZSBzLmFnZSA9IDI3",
+          "Expression" => "c2VsZWN0ICogZnJvbSBvc3NvYmplY3Q=",
           "InputSerialization" => %{"JSON" => %{"Type" => "DOCUMENT"}},
           "OutputSerialization" => %{"JSON" => %{"RecordDelimiter" => "LA=="}}
         }
       }
-      iex> Aliyun.Oss.Object.select_object(config, "some-bucket", "some-object", select_request, format: :json)
+      iex> Aliyun.Oss.Object.select_object(config, "some-bucket", "file.json", select_request, format: :json)
       {:ok, %Aliyun.Oss.Client.Response{
-          data: <<208, 207, 17, 224, 161, 177, 26, 225, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-            0, 0, 0, 62, 0, 3, 0, 254, 255, 9, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18,
-            0, 0, 0, ...>>,
-          headers: [
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          data: %{"contacts" => [...]},
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "connection" => ["keep-alive"],
             ...
-          ]
+          }
         }
       }
       iex> select_request = ~S[
         <SelectRequest>
-            <Expression>c2VsZWN0ICogZnJvbSBvc3NvYmplY3QuY29udGFjdHNbKl0gcyB3aGVyZSBzLmFnZSA9IDI3</Expression>
+            <Expression>c2VsZWN0ICogZnJvbSBvc3NvYmplY3Q=</Expression>
             <InputSerialization>
             <JSON>
                 <Type>DOCUMENT</Type>
@@ -195,15 +224,42 @@ defmodule Aliyun.Oss.Object do
             </OutputSerialization>
         </SelectRequest>
       ]
-      iex> Aliyun.Oss.Object.select_object(config, "some-bucket", "some-object", select_request, format: :json)
+      iex> Aliyun.Oss.Object.select_object(config, "some-bucket", "file.json", select_request, format: :json)
       {:ok, %Aliyun.Oss.Client.Response{
-          data: <<208, 207, 17, 224, 161, 177, 26, 225, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-            0, 0, 0, 62, 0, 3, 0, 254, 255, 9, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18,
-            0, 0, 0, ...>>,
-          headers: [
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          data: %{"contacts" => [...]},
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "connection" => ["keep-alive"],
             ...
-          ]
+          }
+        }
+      }
+      iex> select_request = %{
+        "SelectRequest" => %{"Expression" => "c2VsZWN0ICogZnJvbSBvc3NvYmplY3Q="}
+      }
+      iex> Aliyun.Oss.Object.select_object(config, "some-bucket", "file.csv", select_request, format: :csv)
+      {:ok, %Aliyun.Oss.Client.Response{
+          data: "...",
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "connection" => ["keep-alive"],
+            ...
+          }
+        }
+      }
+      iex> select_request = ~S[
+        <SelectRequest>
+          <Expression>c2VsZWN0ICogZnJvbSBvc3NvYmplY3Q=</Expression>
+        </SelectRequest>
+      ]
+      iex> Aliyun.Oss.Object.select_object(config, "some-bucket", "file.csv", select_request, format: :csv)
+      {:ok, %Aliyun.Oss.Client.Response{
+          data: "...",
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "connection" => ["keep-alive"],
+            ...
+          }
         }
       }
 
@@ -223,7 +279,9 @@ defmodule Aliyun.Oss.Object do
         :json -> "json/select"
       end
 
-    post_object(config, bucket, object, select_request, %{}, %{"x-oss-process" => x_oss_process})
+    post_object(config, bucket, object, select_request,
+      query_params: %{"x-oss-process" => x_oss_process}
+    )
   end
 
   @doc """
@@ -252,10 +310,11 @@ defmodule Aliyun.Oss.Object do
           data: <<208, 207, 17, 224, 161, 177, 26, 225, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
             0, 0, 0, 62, 0, 3, 0, 254, 255, 9, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18,
             0, 0, 0, ...>>,
-          headers: [
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "connection" => ["keep-alive"],
             ...
-          ]
+          }
         }
       }
       iex> select_request = ~S[
@@ -274,10 +333,11 @@ defmodule Aliyun.Oss.Object do
           data: <<208, 207, 17, 224, 161, 177, 26, 225, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
             0, 0, 0, 62, 0, 3, 0, 254, 255, 9, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 18,
             0, 0, 0, ...>>,
-          headers: [
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          headers: %{
+            "accept-ranges" => ["bytes"],
+            "connection" => ["keep-alive"],
             ...
-          ]
+          }
         }
       }
 
@@ -297,36 +357,44 @@ defmodule Aliyun.Oss.Object do
         :json -> "json/meta"
       end
 
-    post_object(config, bucket, object, select_request, %{}, %{"x-oss-process" => x_oss_process})
+    post_object(config, bucket, object, select_request,
+      query_params: %{"x-oss-process" => x_oss_process}
+    )
   end
 
   @doc """
   Creates a signed URL for an object.
 
+  ## Options
+
+  - `:query_params` - Defaults to `%{}`
+  - `:headers` - Defaults to `%{}`
+  - `:method` - HTTP method to use for the signed URL, defaults to `:get`, accept values: `:get`, `:put`, `:post`, `:delete`, `:head`
+  - `:expires` - number of seconds until the signed URL expires, defaults to 3600 seconds (1 hour)
+
   ## Examples
 
-      iex> expires = Timex.now() |> Timex.shift(days: 1) |> Timex.to_unix()
-      iex> Aliyun.Oss.Object.signed_url(config, "some-bucket", "some-object", expires, "GET", %{"Content-Type" => ""})
-      "http://some-bucket.oss-cn-hangzhou.aliyuncs.com/oss-api.pdf?OSSAccessKeyId=nz2pc5*******9l&Expires=1141889120&Signature=vjbyPxybdZ*****************v4%3D"
-      iex> Aliyun.Oss.Object.signed_url(config, "some-bucket", "some-object", expires, "PUT", %{"Content-Type" => "text/plain"})
-      "http://some-bucket.oss-cn-hangzhou.aliyuncs.com/oss-api.pdf?OSSAccessKeyId=nz2pc5*******9l&Expires=1141889120&Signature=vjbyPxybdZ*****************v4%3D"
+      iex> expires = 3600
+      iex> Aliyun.Oss.Object.signed_url(config, "some-bucket", "some-object", expires: expires, method: :get)
+      "https://some-bucket.oss-cn-hangzhou.aliyuncs.com/oss-api.pdf?x-oss-credential=LT**************%2F20250701%2Fcn-hangzhou%2Foss%2Faliyun_v4_request&x-oss-date=20250701T070913Z&x-oss-expires=3600&x-oss-signature=2f64d***********************************************************&x-oss-signature-version=OSS4-HMAC-SHA256"
+      iex> Aliyun.Oss.Object.signed_url(config, "some-bucket", "some-object", expires: expires, method: :put, headers: %{"Content-Type" => "text/plain"})
+      "https://some-bucket.oss-cn-hangzhou.aliyuncs.com/oss-api.pdf?x-oss-credential=LT**************%2F20250701%2Fcn-hangzhou%2Foss%2Faliyun_v4_request&x-oss-date=20250701T070944Z&x-oss-expires=3600&x-oss-signature=2a6e************************************************************&x-oss-signature-version=OSS4-HMAC-SHA256"
 
   """
-  @spec signed_url(Config.t(), String.t(), String.t(), integer(), String.t(), map(), map()) ::
+  @spec signed_url(Config.t(), String.t(), String.t(), keyword()) ::
           String.t()
-  def signed_url(config, bucket, object, expires, method, headers, sub_resources \\ %{}) do
-    request =
-      Request.build(%{
-        verb: method,
-        host: "#{bucket}.#{config.endpoint}",
-        path: "/#{object}",
-        resource: "/#{bucket}/#{object}",
-        sub_resources: sub_resources,
-        headers: headers,
-        expires: expires
-      })
-
-    Request.to_signed_url(config, request)
+  def signed_url(config, bucket, object, options \\ []) do
+    Request.build!(
+      config,
+      Keyword.get(options, :method, :get),
+      bucket,
+      object,
+      Keyword.get(options, :headers, %{}),
+      Keyword.get(options, :query_params, %{}),
+      ""
+    )
+    |> Request.sign_url(Keyword.get(options, :expires, 3600))
+    |> Request.to_url()
   end
 
   @doc """
@@ -334,41 +402,56 @@ defmodule Aliyun.Oss.Object do
 
   ## Examples
 
-      iex> expires = Timex.now() |> Timex.shift(days: 1) |> Timex.to_unix()
+      iex> expires = 3600
       iex> Aliyun.Oss.Object.object_url(config, "some-bucket", "some-object", expires)
       "http://some-bucket.oss-cn-hangzhou.aliyuncs.com/oss-api.pdf?OSSAccessKeyId=nz2pc56s936**9l&Expires=1141889120&Signature=vjbyPxybdZaNmGa%2ByT272YEAiv4%3D"
 
   """
   @spec object_url(Config.t(), String.t(), String.t(), integer()) :: String.t()
   def object_url(config, bucket, object, expires) do
-    signed_url(config, bucket, object, expires, "GET", %{"Content-Type" => ""})
+    signed_url(config, bucket, object, expires: expires, method: :get)
   end
 
   @doc """
   PutObject - uploads objects.
+
+  ## Options
+
+  - `:query_params` - Defaults to `%{}`
+  - `:headers` - Defaults to `%{}`
 
   ## Examples
 
       iex> Aliyun.Oss.Object.put_object(config, "some-bucket", "some-object", "CONTENT")
       {:ok, %Aliyun.Oss.Client.Response{
           data: "",
-          headers: [
-            {"Server", "AliyunOSS"},
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
-            ...
-          ]
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-length" => ["0"],
+            "content-md5" => ["plz3uTkI****************"],
+            "date" => ["Fri, 11 Jul 2025 02:30:40 GMT"],
+            "etag" => ["\"A65CF7B93908B3A*****************\""],
+            "server" => ["AliyunOSS"],
+            "x-oss-hash-crc64ecma" => ["162113910***********"],
+            "x-oss-request-id" => ["687077508A8E************"],
+            "x-oss-server-time" => ["107"]
+          }
         }
       }
 
   """
-  @spec put_object(Config.t(), String.t(), String.t(), String.t(), map(), map()) ::
+  @spec put_object(Config.t(), String.t(), String.t(), String.t(), keyword()) ::
           {:error, error()} | {:ok, Response.t()}
-  def put_object(config, bucket, object, body, headers \\ %{}, sub_resources \\ %{}) do
-    Service.put(config, bucket, object, body, headers: headers, sub_resources: sub_resources)
+  def put_object(config, bucket, object, body, options \\ []) do
+    Service.put(config, bucket, object, body, options)
   end
 
   @doc """
   CopyObject - copies objects within a bucket or between buckets in the same region.
+
+  ## Options
+
+  - `:headers` - Defaults to `%{}`
 
   ## Examples
 
@@ -376,29 +459,38 @@ defmodule Aliyun.Oss.Object do
       {:ok, %Aliyun.Oss.Client.Response{
           data: %{
             "CopyObjectResult" => %{
-              "ETag" => "\"D2D50000000000000000000000000000\"",
-              "LastModified" => "2019-02-27T09:21:13.000Z"
+              "ETag" => "\"D2D5****************************\"",
+              "LastModified" => "2025-02-27T09:21:13.000Z"
             }
           },
-          headers: [
-            {"Server", "AliyunOSS"},
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
-            ...
-          ]
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-type" => ["application/xml"],
+            "date" => ["Fri, 11 Jul 2025 02:35:21 GMT"],
+            "etag" => ["\"A65CF7B9************************\""],
+            "server" => ["AliyunOSS"],
+            "x-oss-copied-size" => ["20"],
+            "x-oss-hash-crc64ecma" => ["16211***************"],
+            "x-oss-ia-retrieve-flow-type" => ["0"],
+            "x-oss-request-id" => ["68707869D***************"],
+            "x-oss-server-time" => ["71"],
+            "x-oss-version-id" => ["null"]
+          }
         }
       }
 
   """
-  @spec copy_object(Config.t(), {String.t(), String.t()}, {String.t(), String.t()}, map()) ::
+  @spec copy_object(Config.t(), {String.t(), String.t()}, {String.t(), String.t()}, keyword()) ::
           {:error, error()} | {:ok, Response.t()}
   def copy_object(
         config,
         {source_bucket, source_object},
         {target_bucket, target_object},
-        headers \\ %{}
+        options \\ []
       ) do
-    headers = Map.put(headers, "x-oss-copy-source", "/#{source_bucket}/#{source_object}")
-    put_object(config, target_bucket, target_object, "", headers)
+    headers = %{"x-oss-copy-source" => "/#{source_bucket}/#{source_object}"}
+    options = Keyword.update(options, :headers, headers, &Map.merge(&1, headers))
+    put_object(config, target_bucket, target_object, "", options)
   end
 
   @doc """
@@ -409,25 +501,27 @@ defmodule Aliyun.Oss.Object do
       iex> Aliyun.Oss.Object.append_object(config, "some-bucket", "some-object", "CONTENT", 0)
       {:ok, %Aliyun.Oss.Client.Response{
           data: "",
-          headers: [
-            {"Server", "AliyunOSS"},
-            {"Date", "Fri, 01 Mar 2019 05:57:23 GMT"},
-            {"Content-Length", "0"},
-            {"Connection", "keep-alive"},
-            {"x-oss-request-id", "5C0000000000000000000000"},
-            {"ETag", "\"B38D0000000000000000000000000000\""},
-            {"x-oss-next-append-position", "10"},
-            {"x-oss-hash-crc64ecma", "8000000000000000000"},
-            {"x-oss-server-time", "17"}
-          ]
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-length" => ["0"],
+            "date" => ["Fri, 11 Jul 2025 02:46:19 GMT"],
+            "etag" => ["\"AE9A6C899***********************\""],
+            "server" => ["AliyunOSS"],
+            "x-oss-hash-crc64ecma" => ["178569**************"],
+            "x-oss-next-append-position" => ["7"],
+            "x-oss-request-id" => ["68707AFB****************"],
+            "x-oss-server-time" => ["15"]
+          }
         }
       }
 
   """
   @spec append_object(Config.t(), String.t(), String.t(), String.t(), integer(), map()) ::
           {:error, error()} | {:ok, Response.t()}
-  def append_object(config, bucket, object, body, position, headers \\ %{}) do
-    post_object(config, bucket, object, body, headers, %{"append" => nil, "position" => position})
+  def append_object(config, bucket, object, content, position, options \\ []) do
+    query_params = %{"append" => nil, "position" => position}
+    options = Keyword.update(options, :query_params, query_params, &Map.merge(&1, query_params))
+    post_object(config, bucket, object, content, options)
   end
 
   @doc """
@@ -438,14 +532,14 @@ defmodule Aliyun.Oss.Object do
       iex> Aliyun.Oss.Object.restore_object(config, "some-bucket", "some-object")
       {:ok, %Aliyun.Oss.Client.Response{
           data: "",
-          headers: [
-            {"Server", "AliyunOSS"},
-            {"Date", "Fri, 01 Mar 2019 06:38:21 GMT"},
-            {"Content-Length", "0"},
-            {"Connection", "keep-alive"},
-            {"x-oss-request-id", "5C7000000000000000000000"},
-            {"x-oss-server-time", "7"}
-          ]
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-length" => ["0"],
+            "date" => ["Fri, 11 Jul 2025 03:13:57 GMT"],
+            "server" => ["AliyunOSS"],
+            "x-oss-request-id" => ["68708175A***************"],
+            "x-oss-server-time" => ["6"]
+          }
         }
       }
 
@@ -453,30 +547,61 @@ defmodule Aliyun.Oss.Object do
   @spec restore_object(Config.t(), String.t(), String.t()) ::
           {:error, error()} | {:ok, Response.t()}
   def restore_object(config, bucket, object) do
-    post_object(config, bucket, object, "", %{}, %{"restore" => nil})
+    post_object(config, bucket, object, "", query_params: %{"restore" => nil})
+  end
+
+  @doc """
+  CleanRestoredObject
+
+  ## Examples
+
+      iex> Aliyun.Oss.Object.clean_restored_object(config, "some-bucket", "some-object")
+      {:ok, %Aliyun.Oss.Client.Response{
+          data: "",
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-length" => ["0"],
+            "date" => ["Fri, 11 Jul 2025 03:13:57 GMT"],
+            "server" => ["AliyunOSS"],
+            "x-oss-request-id" => ["68708175A***************"],
+            "x-oss-server-time" => ["6"]
+          }
+        }
+      }
+
+  """
+  @spec clean_restored_object(Config.t(), String.t(), String.t()) ::
+          {:error, error()} | {:ok, Response.t()}
+  def clean_restored_object(config, bucket, object) do
+    post_object(config, bucket, object, "", query_params: %{"cleanRestoredObject" => nil})
   end
 
   @doc """
   DeleteObject - deletes an object.
+
+  ## Options
+
+  - `:query_params` - Defaults to `%{}`
+  - `:headers` - Defaults to `%{}`
 
   ## Examples
 
       iex> Aliyun.Oss.Object.delete_object(config, "some-bucket", "some-object")
       {:ok, %Aliyun.Oss.Client.Response{
           data: "",
-          headers: [
-            {"Server", "AliyunOSS"},
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-length" => ["0"],
             ...
-          ]
+          }
         }
       }
 
   """
   @spec delete_object(Config.t(), String.t(), String.t()) ::
           {:error, error()} | {:ok, Response.t()}
-  def delete_object(config, bucket, object, sub_resources \\ %{}) do
-    Service.delete(config, bucket, object, sub_resources: sub_resources)
+  def delete_object(config, bucket, object, options \\ []) do
+    Service.delete(config, bucket, object, options)
   end
 
   @doc """
@@ -485,7 +610,7 @@ defmodule Aliyun.Oss.Object do
   ## Options
 
   - `:encoding_type` - Accept value: `:url`
-  - `:quiet` - Set `true` to enable the quiet mode, default is `false`
+  - `:quiet` - Set `true` to enable the quiet mode, default is `false`. This option will be ignored if you pass the raw xml body as the 3rd argument.
 
   ## Examples
 
@@ -493,14 +618,46 @@ defmodule Aliyun.Oss.Object do
       {:ok, %Aliyun.Oss.Client.Response{
           data: %{
             "DeleteResult" => %{
-              "Deleted" => [%{"key" => "object1"}, %{"key" => "object2"}]
+              "Deleted" => [
+                %{"Key" => "object1"},
+                %{"Key" => "object2"}
+              ]
             }
           },
-          headers: [
-            {"Server", "AliyunOSS"},
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-type" => ["application/xml"],
             ...
-          ]
+          }
+        }
+      }
+      iex> xml_body = ~S[
+      <?xml version="1.0" encoding="UTF-8"?>
+      <Delete>
+        <Quiet>false</Quiet>
+        <Object>
+          <Key>multipart.data</Key>
+          <VersionId>CAEQNRiBgIDyz.6C0BYiIGQ2NWEwNmVhNTA3ZTQ3MzM5ODliYjM1ZTdjYjA4****</VersionId>
+        </Object>
+      </Delete>
+      ]
+      iex> Aliyun.Oss.Object.delete_multiple_objects(config, "some-bucket", xml_body)
+      {:ok, %Aliyun.Oss.Client.Response{
+          data: %{
+            "DeleteResult" => %{
+              "Deleted" => [
+                %{
+                  "Key" => "multipart.data",
+                  "VersionId" => "CAEQNRiBgIDyz.6C0BYiIGQ2NWEwNmVhNTA3ZTQ3MzM5ODliYjM1ZTdjYjA4****"
+                },
+              ]
+            }
+          },
+          headers: %{
+            "connection" => ["keep-alive"],
+            "content-type" => ["application/xml"],
+            ...
+          }
         }
       }
   """
@@ -513,22 +670,26 @@ defmodule Aliyun.Oss.Object do
     <% end %>
   </Delete>
   """
-  @spec delete_multiple_objects(Config.t(), String.t(), [String.t()],
-          encoding_type: :url,
-          quiet: boolean()
-        ) ::
+  @spec delete_multiple_objects(Config.t(), String.t(), [String.t()] | String.t(), keyword()) ::
           {:error, error()} | {:ok, Response.t()}
-  def delete_multiple_objects(config, bucket, objects, options \\ []) do
+  def delete_multiple_objects(config, bucket, objects_or_xml_body, options \\ [])
+
+  def delete_multiple_objects(config, bucket, objects, options) when is_list(objects) do
+    quiet = Keyword.get(options, :quiet, false)
+    body = EEx.eval_string(@body_tmpl, quiet: quiet, objects: objects)
+    delete_multiple_objects(config, bucket, body, options)
+  end
+
+  def delete_multiple_objects(config, bucket, body, options) when is_binary(body) do
     headers =
-      case options[:encoding_type] do
+      case Keyword.get(options, :encoding_type) do
         :url -> %{"encoding-type" => "url"}
         _ -> %{}
       end
+      |> Map.put("content-length", String.length(body))
+      |> Map.put("content-md5", :md5 |> :crypto.hash(body) |> Base.encode64())
 
-    quiet = Keyword.get(options, :quiet, false)
-    body = EEx.eval_string(@body_tmpl, quiet: quiet, objects: objects)
-
-    Service.post(config, bucket, nil, body, headers: headers, sub_resources: %{"delete" => nil})
+    Service.post(config, bucket, nil, body, headers: headers, query_params: %{"delete" => nil})
   end
 
   @doc """
@@ -548,7 +709,7 @@ defmodule Aliyun.Oss.Object do
       iex> Aliyun.Oss.Object.sign_post_policy(config, policy)
       %{
         policy: "eyJjb25kaXRpb25zIjpbWyJjb250ZW50LWxlbmd0aC1yYW5nZSIsMCwxMDQ4NTc2MF0seyJidWNrZXQiOiJhaGFoYSJ9LHsiQSI6ImEifSx7ImtleSI6IkFCQyJ9XSwiZXhwaXJhdGlvbiI6IjIwMTMtMTItMDFUMTI6MDA6MDBaIn0=",
-        signature: "W835KpLsL6k1/oo28RcsEflB6hw="
+        signature: "d9ea2e7088a5a7189d2d1a84aa872a00b7078877ffbd4a24b8897c23f16bc1db"
       }
 
   """
@@ -557,16 +718,16 @@ defmodule Aliyun.Oss.Object do
           signature: String.t()
         }
   def sign_post_policy(config, %{} = policy) do
-    secret = config.access_key_secret
-    encoded_policy = policy |> Jason.encode!() |> Base.encode64()
+    today = Date.utc_today() |> Date.to_iso8601(:basic)
+    encoded_policy = policy |> JSON.encode!() |> Base.encode64()
 
     %{
       policy: encoded_policy,
-      signature: Aliyun.Util.Sign.sign(encoded_policy, secret)
+      signature: Sign.sign(encoded_policy, Sign.get_signing_key(config, today))
     }
   end
 
-  defp post_object(config, bucket, object, body, headers, sub_resources) do
-    Service.post(config, bucket, object, body, headers: headers, sub_resources: sub_resources)
+  defp post_object(config, bucket, object, body, options) do
+    Service.post(config, bucket, object, body, options)
   end
 end