+++ test/aliyun_oss/client/request_test.exs
@@ -13,27 +13,19 @@ defmodule Aliyun.Oss.Client.RequestTest do
   @object "exampleobject"
   @region "cn-hangzhou"
   @endpoint "oss-#{@region}.aliyuncs.com"
-  # official example, maybe wrong
-  # @signing_key "3543b7686e65eda71e5e5ca19d548d78423c37e8ddba4dc9d83f90228b457c76"
-  # @signature "053edbf550ebd239b32a9cdfd93b0b2b3f2d223083aa61f75e9ac16856d61f23"
 
   # @signing_key "8a01ff4efcc65ca2cbc75375045c61ab5f3fa8b9a2d84f0add27ef16a25feb3c"
   @signature "d3694c2dfc5371ee6acd35e88c4871ac95a7ba01d3a2f476768fe61218590097"
 
-  setup do
-    config =
-      Config.new!(%{
-        region: @region,
-        endpoint: @endpoint,
-        access_key_id: @access_key_id,
-        access_key_secret: @access_key_secret
-      })
-
-    %{config: config}
-  end
+  @config Config.new!(%{
+            region: @region,
+            endpoint: @endpoint,
+            access_key_id: @access_key_id,
+            access_key_secret: @access_key_secret
+          })
 
   describe "sign with header" do
-    test "creates a request and sign with header", %{config: config} do
+    test "creates a request and sign with header" do
       headers = %{
         "Content-Disposition" => "attachment",
         "content-length" => "3",
@@ -53,7 +45,7 @@ defmodule Aliyun.Oss.Client.RequestTest do
                  }
                }
              } =
-               Request.build!(config, :put, @bucket, @object, headers, %{})
+               Request.build!(@config, :put, @bucket, @object, headers, %{})
                |> Request.sign_header()
     end
   end
@@ -64,12 +56,11 @@ defmodule Aliyun.Oss.Client.RequestTest do
   @timestamp "20250701T024410Z"
   describe "sign with url" do
     test_with_mock "creates a request and sign with url",
-                   %{config: config},
                    DateTime,
                    [:passthrough],
                    utc_now: fn :second -> @utc_now end do
       url =
-        config
+        @config
         |> Request.build!(:get, @bucket, @object, %{}, %{})
         |> Request.sign_url()
         |> Request.to_url()